update tests that try to compare BigInt against BN
`spl-token` field's `getAccount().amount` is now a BigInt, instead of a BN. Due to this, some tests needed to be updated as there is no implicit equality comparison between BN and BigInt.

The only way I found to compare them was to first turn the BN number to string and then initialise a new BigInt from it. It's kinda ugly but I couldn't find a better way.